backend/requirements.txt
alembic==1.17.2 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anthropic==0.18.1 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.12.0 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==5.0.1 ; python_version == "3.11"
asyncpg==0.29.0 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==5.0.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
click==8.3.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==46.0.3 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.19.1 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.110.3 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.20.1 ; python_version >= "3.11" and python_version < "4.0"
fsspec==2025.12.0 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.3.0 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
hf-xet==1.2.0 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "arm64" or platform_machine == "aarch64")
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.26.0 ; python_version >= "3.11" and python_version < "4.0"
huggingface-hub==1.2.3 ; python_version >= "3.11" and python_version < "4.0"
idna==3.11 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
mako==1.3.10 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.3 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.11.5 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.11" and python_version < "4.0"
psycopg-binary==3.3.2 ; python_version >= "3.11" and python_version < "4.0" and implementation_name != "pypy"
psycopg-pool==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy" and implementation_name != "PyPy"
pydantic-core==2.41.5 ; python_version >= "3.11" and python_version < "4.0"
pydantic-extra-types==2.10.6 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.12.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
python-jose==3.5.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.7 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.11" and python_version < "4.0"
redis==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.11" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.45 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.11" and python_version < "4.0"
tokenizers==0.22.1 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
typer-slim==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
ujson==5.11.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.27.1 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.22.1 ; python_version >= "3.11" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.11" and python_version < "4.0"
